chore: begin extracting multicodec utils to global utils
